added integration tests for ordering a service

+++ Main/SpringExample/demo-spring-app-master/src/main/java/BusinessLogicLayer/RestfulObjects/OrderShipmentDto.java
@@ -44,8 +44,8 @@ public class OrderShipmentDto
     public OrderShipmentDto(ShipType shipType,
                             StevedoreServicesOrdered stevedoreServices,
                             boolean usingTeuContainers,
-                            double numberOfContainers,
-                            double numberOfPassengers,
+                            int numberOfContainers,
+                            int numberOfPassengers,
                             String dayOfArrival)
     {
         this.shipType = shipType;
